Added issue notes, default values, and bug fixes
Added issue notes so that once done you can copy and paste the contents
into the issue that you are working on. Also fixed the calculations for
the suggested values since this was off due to a copy paste error.
Finally set default values for the form when you start the application.